backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
httpx==0.25.2
openai==1.98.0
anthropic==0.60.0
google-genai==1.28.2
pymilvus==2.3.4
sqlalchemy==2.0.23
alembic==1.13.1
redis==5.0.1
websockets==12.0
python-dotenv==1.0.0